usbd: device_next: cdc: add option to enable short packet terminate
For usbser.sys driver, which is the default USB host driver for CDC
devices in windows, it is expected that USB device always indicate
completion of transmission by short packet. In case where the last
packet length is multiple of max packet size of the BULK IN endpoint,
the USB device shall indicate completion of transmission by sending a
zero length packet. This commit adds an option to usbd_cdc_acm to
enable the enforcement of the short packet termination mentioned above,
ensuring condition of short packet terminate is fulfilled.

Signed-off-by: Chew Zeh Yang <zeon.chew@ambiq.com>