cmake/modules/yaml: add support for generator expressions
This commit adds support for generator expressions in values and lists
to the yaml module.

The generator expressions are expanded by cmake at the end of the
configuration stage, when the build system is generated; at that moment
the structure of the YAML file would be already defined, and all items
inside a list would get expanded into a single element - definitely not
the desired behavior.

To overcome this limitation, in YAML contexts that use generator
expressions, lists are not directly expanded into proper JSON lists but
stored in their cmake string format with a special marker.
When the yaml_generate() function is called, the internal JSON string is
written using file(GENERATE ...) to a temporary file to obtain a fully
expanded JSON object. The temporary file is then processed by a custom
command that reads the intermediate JSON, replaces the marked strings
with the correct list format, and writes the final YAML file.

Note that:
 - YAML contexts that need to expand generator expressions have to
   specify the GENEX keyword to yaml_create(), or they will be stored
   as raw strings.
 - When generator expressions are used, the YAML file must be written
   using yaml_generate() and not yaml_save(). The output file will be
   available only as a pre-build step of the specified target.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>