dts: include: arm: st: usb: Add USB PHY clock rate support for STM32U5xxx
The existing code assumes that the HSE clock is a 16MHz crystal, however
the hardware allows for a list of possible HSE clock values.

Add them here as an enum, configure it for the STM32U595 chipset
(which is the only device I have access to for testing).

This addresses Issue #79825 .

* Rename USBPHYC_SEL -> OTGHS_SEL which matches the definition in the
  stm32u5 CCIPR2 register (RM0456 Rev 5, Section 11.8.47).
* Add a list of possible values to use in the DT bindings directory
* Add a new PHY (st,stm32-otghs-phy) with an enum list matching the
  above list
* Add support in the USB driver for checking the clock-cfg entry
  and compiling in the correct clock rate.
* And also handle an out of enum configuration by failing compilation.

Signed-off-by: Adrian Chadd <adrian.chadd@meta.com>