drivers: sensor: adxl345: Fix odr update bus fault
The adxl345 accelerometer driver needs to store the current output data
rate (odr) for later use in streaming mode, but the value was
incorrectly placed in the constant configuration structure (located in
flash) and updating it caused a bus fault. Rework odr update handling to
place the current value in RAM instead. The initial value from
devicetree is still stored in flash.

Signed-off-by: Maureen Helm <maureen.helm@analog.com>