net: dhcpv4: update DNS address assignment to be optional
The DHCPv4 options always assigns the DNS address if the router provides one through the DHCP callback. There are times a developer may not always want to assign the server from the router, but only manually assign them. Adding a Kconfig option for assigning the DNS address and defaulting to true to not interfere with previously expected functionality.