llext: disable ELF buffer reuse on MPU/MMU builds
This PR adds a new CONFIG_LLEXT_HW_MEMORY_CHECKS hidden option to the
LLEXT Kconfig, which is enabled by default on builds where memory is
managed by an MPU, MMU or RISC-V PMP, and which may impose additional
alignment or access restrictions to the LLEXT memory regions.

This is used to disable the optimization that allows reuse of the ELF
buffer on such builds, since the ELF sections would certainly not be
aligned properly.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>