apa102: fix end frame
end frame is used to supply clock pulses
so that data goes to last LED in the chain.
Thus, it depends on the number of LEDs in
the chain.
Previously, the number of ones sent into
the end frame was hard-coded and limited
the usage of the driver to 64 LEDs in the
strip.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>