drivers: nxp_enet: Add get_config for ipv6 chksum
Previously, ipv6 being disabled was a dependency of the hw acceleration
of the checksums for the nxp enet driver, because this ethernet has an errata
causing icmpv6 checksum to not be supported. Now, there is a new config
type in ethernet api for checksum types, so we can re-enable hardware
acceleration for ipv6 by implementing this type in the get_config api in
this driver.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>