dts: bindings: serial: nrfx_uarte: Add pin-retention property
Add property to nrfx_uarte which indicates that pins must
be retained for the given instance when it is disabled.

Add that property to uart120 instances as those always
require that since pins are routed through CTRLSEL. For
other instances there are exceptions when CTRLSEL is used
(see the specification).

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>