soc/mediatek: HWMv2 normalization
There is resistance to having the top-level board/SOC in this scheme
be an "audio DSP" (the real devices are integrated application
processors running Linux with GPUs, etc...).  So effect a big rename
a-la HWMv2:

+ Board and SOC get the full product name and lose their "_adsp" suffix
+ "adsp" becomes a CPU Cluster underneath the SOC
+ All the various kconfigs get renamed accordingly

The only user-visible impact is that the "board-only" name disappears,
and you have to specify e.g. "mt8196/mt8196/adsp" to west (though you
can elide the duplicate SOC name as use "mt8196//adsp" if desired).

Signed-off-by: Andy Ross <andyross@google.com>