net: ip: Add check for packet socket for offloaded sockets
This commit fixes the size of struct sockaddr_ptr in case of
CONFIG_NET_SOCKETS_PACKET

The issue was seen when using sendto() with a AF_PACKET socket.
in net_context.h, struct net_context contains:
- struct sockaddr_ptr local
- struct sockaddr remote

Without the fix, local is on 16 bytes, but handle data up to 20 bytes.
As a aconsequence, when remote gets updated in context_sendto(), the last
part of local (bytes 15 to 20) gets overwritten.

Signed-off-by: Noemie Gillet <ngillet@sequans.com>